pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegis"
version = "0.1.0"
description = "Aegis - Your Personal AI Assistant Platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "hazo" }]
keywords = ["ai", "assistant", "llm", "personal-assistant", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",

    # LLM Providers
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "ollama>=0.4.0",
    "tiktoken>=0.8.0",

    # Storage
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",

    # Authentication
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Utilities
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "croniter>=2.0.0",
    "pyngrok>=7.0.0",
    "html2text>=2024.2.26",

    # Document export
    "python-docx>=1.1.0",
    "markdown>=3.5.0",
    "weasyprint>=62.0",

    # Data science (for Python interpreter)
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "openpyxl>=3.1.0",
    "scipy>=1.11.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
full = [
    # Memory - semantic search
    "chromadb>=0.5.0",
    # Browser automation
    "playwright>=1.48.0",
    # Task scheduling
    "apscheduler>=3.10.0",
    # Skills
    "python-frontmatter>=1.1.0",
    "watchfiles>=0.24.0",
    # Secrets
    "keyring>=25.5.0",
]

[project.scripts]
aegis = "aegis.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aegis"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
